requirements.txt
# Core ML Libraries
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0
safetensors>=0.4.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Jupyter and Colab
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0

# Text Processing
regex>=2023.8.8
nltk>=3.8.1

# Model Training and Evaluation
datasets>=2.15.0
evaluate>=0.4.0
tensorboard>=2.15.0

# Image Generation (for Stable Diffusion notebook)
diffusers>=0.25.0
gradio>=4.12.0
Pillow>=10.0.0

# API and Authentication
huggingface-hub>=0.19.0
requests>=2.31.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
psutil>=5.9.0

# Memory optimization
bitsandbytes>=0.41.0

# Additional ML utilities
wandb>=0.16.0
peft>=0.7.0
chardet
GitPython
